+++ crates/rome_js_formatter/src/js/bogus/bogus_import_assertion_entry.rs
@@ -1,7 +1,0 @@
-use crate::FormatBogusNodeRule;
-use rome_js_syntax::JsBogusImportAssertionEntry;
-
-#[derive(Debug, Clone, Default)]
-pub(crate) struct FormatJsBogusImportAssertionEntry;
-
-impl FormatBogusNodeRule<JsBogusImportAssertionEntry> for FormatJsBogusImportAssertionEntry {}