+++ crates/rome_js_formatter/src/js/bogus/bogus_import_attribute_entry.rs
@@ -1,0 +1,5 @@
+use crate::FormatBogusNodeRule;
+use rome_js_syntax::JsBogusImportAttributeEntry;
+#[derive(Debug, Clone, Default)]
+pub(crate) struct FormatJsBogusImportAttributeEntry;
+impl FormatBogusNodeRule<JsBogusImportAttributeEntry> for FormatJsBogusImportAttributeEntry {}