+++ crates/rome_js_formatter/tests/specs/js/module/export/named_from_clause.js
@@ -1,20 +1,20 @@
 export {a,
     b as c
-} from "fancy" assert { type: "json"}
+} from "fancy" with { type: "json"}
 
 
 export {
     lorem_lorem_lorem_lorem_lorem_lorem_lorem_lorem_lorem,
     lorem_lorem_lorem_lorem_lorem_ as ipsum_ipsum_ipsum_ipsum_ipsum_ipsum_
-} from "fancy" assert { type: "json", "type2": "json", type23: "json", "type24": "json"}
+} from "fancy" with { type: "json", "type2": "json", type23: "json", "type24": "json"}
 
 
 export { loooooooooooooooooooooooooooooooooooooooooooooooooong } from "loooooooooooooooooooooooooooooooooooooooooooooong"
 
 export { loooooooooooooooooooooooooooooooooooooooooooooooooong, } from "loooooooooooooooooooooooooooooooooooooooooooooong"
 
-export { 
-// comment 
+export {
+// comment
 loooooooooooooooooooooooooooooooooooooooooooooooooong } from "loooooooooooooooooooooooooooooooooooooooooooooong"
 
 export { a as b } from "loooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong"