+++ crates/rome_js_parser/src/lexer/tests.rs
@@ -1311,7 +1311,6 @@ fn keywords() {
         "abstract",
         "as",
         "asserts",
-        "assert",
         "any",
         "async",
         "await",