+++ crates/rome_js_parser/src/syntax/module.rs
@@ -248,14 +248,14 @@ fn parse_import_default_or_named_clause_rest(
 
             p.expect(T![from]);
             parse_module_source(p).or_add_diagnostic(p, expected_module_source);
-            parse_import_assertion(p).ok();
+            parse_import_attributes(p).ok();
 
             named_clause.complete(p, JS_IMPORT_NAMED_CLAUSE)
         }
         _ => {
             p.expect(T![from]);
             parse_module_source(p).or_add_diagnostic(p, expected_module_source);
-            parse_import_assertion(p).ok();
+            parse_import_attributes(p).ok();
 
             m.complete(p, JS_IMPORT_DEFAULT_CLAUSE)
         }
@@ -268,7 +268,7 @@ fn parse_import_default_or_named_clause_rest(
 fn parse_import_bare_clause(p: &mut JsParser) -> ParsedSyntax {
     parse_module_source(p).map(|module_source| {
         let m = module_source.precede(p);
-        parse_import_assertion(p).ok();
+        parse_import_attributes(p).ok();
         m.complete(p, JS_IMPORT_BARE_CLAUSE)
     })
 }
@@ -282,7 +282,7 @@ fn parse_import_namespace_clause_rest(p: &mut JsParser, m: Marker) -> CompletedM
     parse_binding(p).or_add_diagnostic(p, expected_binding);
     p.expect(T![from]);
     parse_module_source(p).or_add_diagnostic(p, expected_module_source);
-    parse_import_assertion(p).ok();
+    parse_import_attributes(p).ok();
 
     m.complete(p, JS_IMPORT_NAMESPACE_CLAUSE)
 }
@@ -298,7 +298,7 @@ fn parse_import_named_clause_rest(p: &mut JsParser, m: Marker) -> CompletedMarke
     parse_named_import(p).or_add_diagnostic(p, expected_named_import);
     p.expect(T![from]);
     parse_module_source(p).or_add_diagnostic(p, expected_module_source);
-    parse_import_assertion(p).ok();
+    parse_import_attributes(p).ok();
 
     m.complete(p, JS_IMPORT_NAMED_CLAUSE)
 }
@@ -445,51 +445,51 @@ fn parse_any_named_import_specifier(p: &mut JsParser) -> ParsedSyntax {
     }
 }
 
-// test import_assertion
-// import "x" assert { type: "json" }
-// import "foo" assert { "type": "json" };
-// import foo from "foo.json" assert { type: "json" };
-// import {test} from "foo.json" assert { for: "for" }
-// import foo_json from "foo.json" assert { type: "json", hasOwnProperty: "true" };
-// import "x" assert
+// test import_attribute
+// import "x" with { type: "json" }
+// import "foo" with { "type": "json" };
+// import foo from "foo.json" with { type: "json" };
+// import {test} from "foo.json" with { for: "for" }
+// import foo_json from "foo.json" with { type: "json", hasOwnProperty: "true" };
+// import "x" with
 // { type: "json" }
 
-// test_err import_assertion_err
-// import "foo" assert { type, "json" };
-// import "bar" \u{61}ssert { type: "json" };
-// import { foo } assert { type: "json" };
+// test_err import_attribute_err
+// import "foo" with { type, "json" };
+// import "bar" \u{61}ith { type: "json" };
+// import { foo } with { type: "json" };
 // import "lorem"
 // assert { type: "json" }
-// import foo2 from "foo.json" assert { "type": "json", type: "html", "type": "js" };
-// import "x" assert;
-// import ipsum from "ipsum.json" assert { type: "json", lazy: true, startAtLine: 1 };
-// import { a } from "a.json" assert
-fn parse_import_assertion(p: &mut JsParser) -> ParsedSyntax {
-    if !p.at(T![assert]) || p.has_preceding_line_break() {
+// import foo2 from "foo.json" with { "type": "json", type: "html", "type": "js" };
+// import "x" with;
+// import ipsum from "ipsum.json" with { type: "json", lazy: true, startAtLine: 1 };
+// import { a } from "a.json" with
+fn parse_import_attributes(p: &mut JsParser) -> ParsedSyntax {
+    if !p.at(T![with]) || p.has_preceding_line_break() {
         return Absent;
     }
 
     let m = p.start();
-    p.expect(T![assert]);
+    p.expect(T![with]);
     p.expect(T!['{']);
 
-    ImportAssertionList::default().parse_list(p);
+    ImportAttributeList::default().parse_list(p);
 
     p.expect(T!['}']);
 
-    Present(m.complete(p, JS_IMPORT_ASSERTION))
+    Present(m.complete(p, JS_IMPORT_ATTRIBUTE))
 }
 
 #[derive(Default)]
-struct ImportAssertionList {
+struct ImportAttributeList {
     assertion_keys: HashMap<String, TextRange>,
 }
 
-impl ParseSeparatedList for ImportAssertionList {
+impl ParseSeparatedList for ImportAttributeList {
     type Kind = JsSyntaxKind;
     type Parser<'source> = JsParser<'source>;
 
-    const LIST_KIND: Self::Kind = JS_IMPORT_ASSERTION_ENTRY_LIST;
+    const LIST_KIND: Self::Kind = JS_IMPORT_ATTRIBUTE_ENTRY_LIST;
 
     fn parse_element(&mut self, p: &mut JsParser) -> ParsedSyntax {
         parse_import_assertion_entry(p, &mut self.assertion_keys)
@@ -503,7 +503,7 @@ impl ParseSeparatedList for ImportAssertionList {
         parsed_element.or_recover(
             p,
             &ParseRecovery::new(
-                JS_BOGUS_IMPORT_ASSERTION_ENTRY,
+                JS_BOGUS_IMPORT_ATTRIBUTE_ENTRY,
                 STMT_RECOVERY_SET.union(token_set![T![,], T!['}']]),
             )
             .enable_recovery_on_line_break(),
@@ -572,7 +572,7 @@ fn parse_import_assertion_entry(
     p.expect(T![:]);
     p.expect(JS_STRING_LITERAL);
 
-    let mut entry = m.complete(p, JS_IMPORT_ASSERTION_ENTRY);
+    let mut entry = m.complete(p, JS_IMPORT_ATTRIBUTE_ENTRY);
 
     if !valid {
         entry.change_to_bogus(p);
@@ -917,7 +917,7 @@ where
 // export * from "a";
 // export * as c from "b";
 // export * as default from "b"
-// export * from "mod" assert { type: "json" }
+// export * from "mod" with { type: "json" }
 // export type * from "types";
 // export type * as types from "types";
 //
@@ -944,7 +944,7 @@ fn parse_export_from_clause(p: &mut JsParser) -> ParsedSyntax {
     parse_export_as_clause(p).ok();
     p.expect(T![from]);
     parse_module_source(p).or_add_diagnostic(p, expected_module_source);
-    parse_import_assertion(p).ok();
+    parse_import_attributes(p).ok();
     semi(p, TextRange::new(start, p.cur_range().end()));
 
     Present(m.complete(p, JS_EXPORT_FROM_CLAUSE))
@@ -956,7 +956,7 @@ fn parse_export_from_clause(p: &mut JsParser) -> ParsedSyntax {
 // export { as } from "mod";
 // export { default as "b" } from "mod";
 // export { "a" as b } from "mod";
-// export { a } from "mod" assert { type: "json" }
+// export { a } from "mod" with { type: "json" }
 // export { "a" } from "./mod";
 // export {
 //      "a"
@@ -990,7 +990,7 @@ fn parse_export_named_from_clause(p: &mut JsParser) -> ParsedSyntax {
     p.expect(T![from]);
 
     parse_module_source(p).or_add_diagnostic(p, expected_module_source);
-    parse_import_assertion(p).ok();
+    parse_import_attributes(p).ok();
 
     semi(p, TextRange::new(start, p.cur_range().start()));