+++ crates/rome_js_parser/test_data/inline/err/import_attribute_err.js
@@ -1,0 +1,9 @@
+import "foo" with { type, "json" };
+import "bar" \u{61}ith { type: "json" };
+import { foo } with { type: "json" };
+import "lorem"
+assert { type: "json" }
+import foo2 from "foo.json" with { "type": "json", type: "html", "type": "js" };
+import "x" with;
+import ipsum from "ipsum.json" with { type: "json", lazy: true, startAtLine: 1 };
+import { a } from "a.json" with